requirements.txt
gradio
Pillow
pandas
piexif
